+++ examples/query-complexity/recipe-resolver.ts
@@ -1,4 +1,4 @@
-import { Resolver, Query, FieldResolver, Arg, Root, Int, ResolverInterface } from "../../src";
+import { Resolver, Query, FieldResolver, Root, ResolverInterface, Arg } from "../../src";
 
 import { Recipe } from "./recipe-type";
 import { createRecipeSamples } from "./recipe-samples";
@@ -7,16 +7,23 @@ import { createRecipeSamples } from "./recipe-samples";
 export class RecipeResolver implements ResolverInterface<Recipe> {
   private readonly items: Recipe[] = createRecipeSamples();
 
-  @Query(returns => [Recipe], { description: "Get all the recipes from around the world " })
-  async recipes(): Promise<Recipe[]> {
-    return await this.items;
+  @Query(returns => [Recipe], {
+    /*
+      You can also pass a calculation function in the complexity option
+      to determine a custom complexity. This function will provide the
+      complexity of the child nodes as well as the field input arguments.
+      That way you can make a more realistic estimation of individual field
+      complexity values, e.g. by multiplying childComplexity by the number of items in array
+    */
+    complexity: ({ childComplexity, args }) => args.count * childComplexity,
+  })
+  async recipes(@Arg("count") count: number): Promise<Recipe[]> {
+    return await this.items.slice(0, count);
   }
-  /* Complexity in field resolver overrides complexity of equivalent field type*/
-  @FieldResolver({ complexity: 10 })
-  ratingsCount(
-    @Root() recipe: Recipe,
-    @Arg("minRate", type => Int, { nullable: true }) minRate: number = 0.0,
-  ): number {
-    return recipe.ratings.filter(rating => rating >= minRate).length;
+
+  /* Complexity in field resolver overrides complexity of equivalent field type */
+  @FieldResolver({ complexity: 5 })
+  ratingsCount(@Root() recipe: Recipe): number {
+    return recipe.ratings.length;
   }
 }