+++ tests/helpers/getSchemaInfo.ts
@@ -1,9 +1,8 @@
 import {
   graphql,
-  introspectionQuery,
+  getIntrospectionQuery,
   IntrospectionObjectType,
   IntrospectionSchema,
-  GraphQLSchema,
 } from "graphql";
 
 import { buildSchema, BuildSchemaOptions } from "../../src";
@@ -13,7 +12,7 @@ export async function getSchemaInfo(options: BuildSchemaOptions) {
   const schema = await buildSchema(options);
 
   // get builded schema info from retrospection
-  const result = await graphql(schema, introspectionQuery);
+  const result = await graphql(schema, getIntrospectionQuery());
   expect(result.errors).toBeUndefined();
 
   const schemaIntrospection = result.data!.__schema as IntrospectionSchema;