+++ tests/helpers/subscriptions/createWebSocketGQL.ts
@@ -1,39 +1,36 @@
 import * as NodeWebSocket from "ws";
-import { AddressInfo } from "net";
-import { Server as HTTPServer } from "http";
-import { Server as HTTPSServer } from "https";
 import { GraphQLSchema } from "graphql";
-import { GraphQLServer } from "graphql-yoga";
+import { ApolloServer } from "apollo-server";
 import { SubscriptionClient } from "subscriptions-transport-ws";
 import { ApolloClient } from "apollo-client";
 import { WebSocketLink } from "apollo-link-ws";
 import { InMemoryCache } from "apollo-cache-inmemory";
 
 export interface WebSocketUtils {
-  apollo: ApolloClient<any>;
-  server: HTTPServer | HTTPSServer;
+  apolloClient: ApolloClient<unknown>;
+  apolloServer: ApolloServer;
 }
 
-export async function createWebSocketUtils(schema: GraphQLSchema): Promise<WebSocketUtils> {
-  const graphQLServer = new GraphQLServer({ schema });
-  const server = await graphQLServer.start({
-    port: 0,
+export default async function createWebSocketUtils(schema: GraphQLSchema): Promise<WebSocketUtils> {
+  const apolloServer = new ApolloServer({
+    schema,
     playground: false,
-    endpoint: "/graphql",
-    subscriptions: "/graphql",
+  });
+  const { subscriptionsUrl } = await apolloServer.listen({
+    port: 0,
   });
 
   const subscriptionClient = new SubscriptionClient(
-    `ws://localhost:${(server.address() as AddressInfo).port}/graphql`,
+    subscriptionsUrl,
     { reconnect: true },
     NodeWebSocket,
   );
-  const apollo = new ApolloClient({
+  const apolloClient = new ApolloClient({
     link: new WebSocketLink(subscriptionClient),
     cache: new InMemoryCache(),
   });
   return {
-    apollo,
-    server,
+    apolloClient,
+    apolloServer,
   };
 }